+++ packages/ast-utilities/loom.config.ts
@@ -1,11 +1,10 @@
-import {createPackage, Runtime} from '@shopify/loom';
+import {createPackage} from '@shopify/loom';
 
 import {quiltPackage} from '../../config/loom';
 
 export default createPackage((pkg) => {
-  pkg.runtimes(Runtime.Browser, Runtime.Node);
-  pkg.entry({root: './src/index'});
-  pkg.entry({name: 'javascript', root: './src/javascript'});
-  pkg.entry({name: 'markdown', root: './src/markdown'});
+  pkg.entry({root: './src/index.ts'});
+  pkg.entry({name: 'javascript', root: './src/javascript/index.ts'});
+  pkg.entry({name: 'markdown', root: './src/markdown/index.ts'});
   pkg.use(quiltPackage());
 });