+++ packages/graphql-validate-fixtures/loom.config.ts
@@ -1,10 +1,9 @@
-import {createPackage, Runtime} from '@shopify/loom';
+import {createPackage} from '@shopify/loom';
 
 import {quiltPackage} from '../../config/loom';
 
 export default createPackage((pkg) => {
-  pkg.runtimes(Runtime.Browser, Runtime.Node);
-  pkg.entry({root: './src/index'});
-  pkg.binary({name: 'graphql-validate-fixtures', root: './src/cli'});
+  pkg.entry({root: './src/index.ts'});
+  pkg.binary({name: 'graphql-validate-fixtures', root: './src/cli.ts'});
   pkg.use(quiltPackage());
 });