+++ packages/web-worker/loom.config.ts
@@ -3,11 +3,10 @@ import {createPackage, createProjectBuildPlugin, Runtime} from '@shopify/loom';
 import {quiltPackage} from '../../config/loom';
 
 export default createPackage((pkg) => {
-  pkg.runtimes(Runtime.Browser, Runtime.Node);
-  pkg.entry({root: './src/index'});
-  pkg.entry({name: 'babel', root: './src/babel-plugin'});
-  pkg.entry({name: 'webpack', root: './src/webpack-parts'});
-  pkg.entry({name: 'worker', root: './src/worker'});
+  pkg.entry({root: './src/index.ts'});
+  pkg.entry({name: 'babel', root: './src/babel-plugin.ts'});
+  pkg.entry({name: 'webpack', root: './src/webpack-parts/index.ts'});
+  pkg.entry({name: 'worker', root: './src/worker.ts'});
   pkg.use(quiltPackage());
 
   // The library references the worker "wrapper" files as relative paths from