+++ packages/web-worker/CHANGELOG.md
@@ -7,6 +7,16 @@ and adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 <!-- ## Unreleased -->
 
+## 4.0.2 - 2022-02-01
+
+### Changed
+
+- Rerelease after failed publish. No code changes.
+
+## 4.0.1 - 2022-02-01
+
+- No updates. Transitive dependency bump.
+
 ## 4.0.0 - 2021-12-01
 
 ### Breaking Change